Update changes as suggested by the code reviewer
