+++ my_list.rb
@@ -1,0 +1,14 @@
+require_relative 'my_enumerable'
+
+class MyList
+  include MyEnumerable
+  def initialize(*elements)
+    @list = elements
+  end
+
+  def each
+    @list.each do |element|
+      yield element
+    end
+  end
+end
